requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
tensorflow==2.18.0
numpy>=1.26.0,<2.1.0
Pillow==9.5.0
python-multipart==0.0.5
firebase-admin==6.0.0
